% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukpid_libraries.R
\name{ukpid_libraries}
\alias{ukpid_libraries}
\title{Load and Install Libraries}
\usage{
ukpid_libraries(
  packages = c("tidyverse", "ggplot2", "readxl", "lubridate", "scico", "knitr", "dplyr",
    "data.table", "formattable", "scales", "gridExtra", "openxlsx", "eeptools",
    "stringr", "gt", "janitor", "tidyquant", "patchwork", "survival", "survminer",
    "ggsurvfit", "plotly")
)
}
\arguments{
\item{packages}{A character vector of library names to load. Default includes "ggplot2", "dplyr", "tidyr", "readr".}
}
\description{
This function installs and loads a predefined set of libraries or user-specified ones.
}
\examples{
ukpid_libraries()
ukpid_libraries(c("data.table", "lubridate"))
}
